Restore tests not related to a contract
